+++ src/frontend/typer/fagsakdeltager.ts
@@ -3,6 +3,7 @@ import { kjønnType } from '@navikt/familie-typer';
 export enum FagsakDeltagerRolle {
     Barn = 'BARN',
     Forelder = 'FORELDER',
+    Ukjent = 'UKJENT',
 }
 
 export interface IFagsakDeltager {
@@ -10,5 +11,9 @@ export interface IFagsakDeltager {
     ident: string;
     rolle: FagsakDeltagerRolle;
     kjønn?: kjønnType;
-    fagsakId?: string;
+    fagsakId?: number;
+}
+
+export interface ISøkParam {
+    personIdent: string;
 }
